% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustImmuneLENS.R
\name{adjustImmuneLENS}
\alias{adjustImmuneLENS}
\title{Adjust TCRA T cell fraction based on tumour purity and copy number}
\usage{
adjustImmuneLENS(out.df, purity, local.cn, vdj.gene)
}
\arguments{
\item{out.df}{Output containing unadjusted T/B cell fraction}

\item{purity}{Tumour purity, between 0 and 1}

\item{local.cn}{Tumour copynumber at VDJ gene locus}

\item{vdj.gene}{VDJ gene of interest}
}
\value{
data frame of TCRA scores with purity/copy number adjusted outputs
}
\description{
Adjust TCRA T cell fraction based on tumour purity and copy number
}
