% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_fasta}
\alias{all_fasta}
\title{FASTA sequences for VDJ genes (hg19)}
\format{
a list containing SegFastadna objects
}
\usage{
all_fasta
}
\value{
NA
}
\description{
FAST Sequences
}
\keyword{datasets}
