% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_fasta_hg38}
\alias{all_fasta_hg38}
\title{FASTA sequences for VDJ genes (hg38)}
\format{
a list containing SegFastadna objects
}
\usage{
all_fasta_hg38
}
\value{
NA
}
\description{
FAST Sequences
}
\keyword{datasets}
