% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_haplotype_regions_de_novo_diploid.R
\name{calc_haplotype_regions_de_novo_diploid}
\alias{calc_haplotype_regions_de_novo_diploid}
\title{Calculate predicted IGH germline haplotype assuming low B cell content + diploid sample}
\usage{
calc_haplotype_regions_de_novo_diploid(
  input_cov,
  kb_len_threshold = 5,
  bin_size = 1000
)
}
\arguments{
\item{input_cov}{data frame of positions and coverage values for IGH loci}

\item{kb_len_threshold}{threshold for length of genomic regions used in normalisation method (default = 5kb)}

\item{bin_size}{default = 1000}
}
\value{
regions used for later normalisation
}
\description{
Calculate predicted IGH germline haplotype assuming low B cell content + diploid sample
}
