% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcJSD_distance.R
\name{calcJSD_distance}
\alias{calcJSD_distance}
\title{Function to calculate distance in V or J segment usage between two samples}
\usage{
calcJSD_distance(out_df1, out_df2, V_or_J)
}
\arguments{
\item{out_df1}{data frame output from runTcellExtrect_WGS containing segment fractions for sample 1}

\item{out_df2}{data frame output from runTcellExtrect_WGS containing segment fractions for sample 2}

\item{V_or_J}{plot V or J segment proportions}
}
\value{
JSD_dist
}
\description{
Function to calculate distance in V or J segment usage between two samples
}
