% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exonsImmuneLENS.R
\name{exonsImmuneLENS}
\alias{exonsImmuneLENS}
\title{Function to run ImmuneLENS for exon output}
\usage{
exonsImmuneLENS(
  vdj.region.df,
  vdj.gene = "TCRA",
  hg19_or_38 = "hg38",
  GC_correct = TRUE,
  median.k = 50,
  median.thresh = 15,
  sample_name = "test",
  customSeg = NULL,
  customFasta = NULL,
  customFlag = NULL,
  removed_flag = TRUE
)
}
\arguments{
\item{vdj.region.df}{data frame containing coverage values by position}

\item{vdj.gene}{VDJ gene to use}

\item{hg19_or_38}{hg19 or hg38 version of genome}

\item{GC_correct}{whether to use GC correction or not for output}

\item{median.k}{rolling median window}

\item{median.thresh}{threshold to remove exons with low coverage}

\item{sample_name}{name of sample run}

\item{customSeg}{custom seg file}

\item{customFasta}{custom Fasta file}

\item{customFlag}{custom list of exons to flag for removal}

\item{removed_flag}{Whether to removed flagged WGS locations (default = TRUE)}
}
\value{
data frame of TCRA T cell fractions with 95\% CI
}
\description{
Function to run ImmuneLENS for exon output
}
