% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCcorrect_WGS.R
\name{GCcorrect_WGS}
\alias{GCcorrect_WGS}
\title{Method for correcting for GC biases}
\usage{
GCcorrect_WGS(
  tumour.logR,
  exons,
  exonList,
  gene.fasta,
  gene.fasta.start,
  hg19_or38 = "hg38",
  sliding = 1000
)
}
\arguments{
\item{tumour.logR}{LogR values}

\item{exons}{Location of exons}

\item{exonList}{GC content of exons}

\item{gene.fasta}{offset for FASTA file, e.g 21999999 for TCRA in hg19}

\item{gene.fasta.start}{The chr position the fasta file begins}

\item{hg19_or38}{What genome version to use, must be 'hg19' or 'hg38'}

\item{sliding}{number of bp for gc windows}
}
\value{
data frame of GC correct logR values
}
\description{
Method for correcting for GC biases
}
