% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCovFromBam_WGS.R
\name{getCovFromBam_WGS}
\alias{getCovFromBam_WGS}
\title{Function to create TCRA coverage file from bams}
\usage{
getCovFromBam_WGS(
  bamPath,
  outPath,
  vdj.gene,
  hg19_or_38,
  samtools.loc = "",
  index.loc = NULL,
  customSeg = NULL
)
}
\arguments{
\item{bamPath}{Path to bam file}

\item{outPath}{Path to output directory}

\item{vdj.gene}{VDJ gene to use}

\item{hg19_or_38}{Human genome version}

\item{samtools.loc}{Path to working version of samtools}

\item{index.loc}{Path to BAM index if in different path to BAM}

\item{customSeg}{Custom segment file}
}
\description{
Function to create TCRA coverage file from bams
}
