% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVDJFraction_ImmuneLENS.R
\name{getVDJfraction_ImmuneLENS}
\alias{getVDJfraction_ImmuneLENS}
\title{Function to run ImmuneLENS for WGS for individual V and J segments}
\usage{
getVDJfraction_ImmuneLENS(
  test.logR,
  vdj.gene,
  sample_name,
  hg19_or_38 = "hg38",
  GC.correct = TRUE,
  restriktor.absval = 1e-05,
  exons = NULL,
  exonList = NULL,
  gene.fasta = NULL,
  gene.fasta.start = NULL,
  sliding = 1000,
  classSwitch = TRUE,
  customNorm = NULL,
  IGH.gc.constraint = TRUE,
  IGH.gc.constraint.value = 0.01,
  TCRB.gc.constraint = TRUE,
  allGC = FALSE
)
}
\arguments{
\item{test.logR}{data frame containing logR values to be fitted to constrained model}

\item{vdj.gene}{V(D)J gene to use}

\item{sample_name}{name of the sample}

\item{hg19_or_38}{version of the genome}

\item{GC.correct}{whether GC.correct is being used}

\item{restriktor.absval}{Parameter for use in Restriktor}

\item{exons}{Location of exons}

\item{exonList}{GC content of exons}

\item{gene.fasta}{FASTA file for VDJ gene e.g TCRA}

\item{gene.fasta.start}{offset for FASTA file, e.g 21999999 for TCRA in hg19}

\item{sliding}{number of bp for gc windows (default = 1000)}

\item{classSwitch}{Whether to include class switching in model for IGH (default = TRUE)}

\item{customNorm}{Custom dataframe with average values to normalise for, summarised over 100bp (default = NULL)}

\item{IGH.gc.constraint}{TRUE/FALSE, use additional constraints for GC correction smooth.gc and smooth.gc2 in the IGH solution (default TRUE to prevent over-fitting)}

\item{IGH.gc.constraint.value}{Value to constrain smooth.gc and smooth.gc2 by (default 0.01)}

\item{TCRB.gc.constraint}{TRUE/FALSE, use additional constraints for GC correction smooth.gc and smooth.gc2 in the TCRB solution (default TRUE to prevent over-fitting)}

\item{allGC}{TRUE/FALSE, when TRUE make model where everything is due to GC}
}
\value{
data frame of TCRA T cell fractions for VDJ segments
}
\description{
Function to run ImmuneLENS for WGS for individual V and J segments
}
