% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGH_gc_correct.R
\name{IGH_gc_correct}
\alias{IGH_gc_correct}
\title{GC correct IGH coverage values for use prior to identification of IGH haplotype}
\usage{
IGH_gc_correct(input_cov, hg19_or38 = "hg38")
}
\arguments{
\item{input_cov}{data frame of positions and coverage values for IGH loci}

\item{hg19_or38}{default = 'hg38'}
}
\value{
Data frame of GC corrected coverage values
}
\description{
GC correct IGH coverage values for use prior to identification of IGH haplotype
}
