% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGH_haplotype_norm_for_region_diploid.R,
%   R/IGH_haplotype_norm_for_region_tumour.R
\name{IGH_haplotype_norm_for_region_diploid}
\alias{IGH_haplotype_norm_for_region_diploid}
\alias{IGH_haplotype_norm_for_region_tumour}
\title{Calculate predicted IGH germline haplotype assuming possible high B cell content}
\usage{
IGH_haplotype_norm_for_region_diploid(input_cov, region_df)

IGH_haplotype_norm_for_region_tumour(
  input_cov,
  region_df,
  norm_type = "benchmark",
  gc_correct = FALSE
)
}
\arguments{
\item{input_cov}{data frame of positions and coverage values for IGH loci}

\item{region_df}{predicted IGH germline haplotype regions}

\item{norm_type}{type of normalisation to use (default = 'benchmark')}

\item{gc_correct}{logical, should the normalisation be GC corrected? (default = FALSE)}
}
\value{
IGH coverage data frame with normalised values by predicted IGH haplotype

IGH coverage data frame with normalised values by predicted IGH germline haplotype applied to a tumour sample
}
\description{
Calculate predicted IGH germline haplotype assuming possible high B cell content

Function to normalise IGH coverage values by predicted IGH germline haplotype regions for a tumour sample
}
