% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGH_haplotype_norm_fun.R
\name{IGH_haplotype_norm_fun}
\alias{IGH_haplotype_norm_fun}
\title{Calculate normalised coverage file for the predicted IGH germline haplotype}
\usage{
IGH_haplotype_norm_fun(
  cov.df,
  kb_len_threshold = 5,
  bin_size = 1000,
  highBcell = FALSE
)
}
\arguments{
\item{cov.df}{data frame of positions and coverage values for IGH loci}

\item{kb_len_threshold}{threshold for length of genomic regions used in normalisation method (default = 5kb)}

\item{bin_size}{bins used for normalisation (default = 1000)}

\item{highBcell}{Is the sample high in B cells (> 0.25 fraction)? (default = FALSE)}
}
\value{
List of 1. normalised coverage dataframe and 2. regions used for normalisation
}
\description{
Calculate normalised coverage file for the predicted IGH germline haplotype
}
