% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGH_haplotype_norm_fun_tumour.R
\name{IGH_haplotype_norm_fun_tumour}
\alias{IGH_haplotype_norm_fun_tumour}
\title{Calculate normalised coverage file for the predicted IGH germline haplotype for a tumour sample}
\usage{
IGH_haplotype_norm_fun_tumour(
  cov.df,
  cov.df.tumour,
  regions_gl_df1,
  gc_correct = TRUE,
  run_breakpoint_check = TRUE,
  threshold_to_norm = 0,
  norm_type_to_use = "benchmark"
)
}
\arguments{
\item{cov.df}{data frame of positions and coverage values for IGH loci of matched germline sample}

\item{cov.df.tumour}{data frame of positions and coverage values for IGH loci of tumour sample}

\item{regions_gl_df1}{data frame of regions used for normalisation from germline sample}

\item{gc_correct}{logical, should the normalisation be GC corrected? (default = TRUE)}

\item{run_breakpoint_check}{logical, should the normalisation be checked for breakpoints? (default = TRUE)}

\item{threshold_to_norm}{threshold for normalisation (default = 0)}

\item{norm_type_to_use}{type of normalisation to use (default = 'benchmark')}
}
\value{
List of four items. 1. normalised coverage without somatic normalisation. 2. normalisation coverage with somatic normalisaton. 3. Somatic regions predicted and used for normalisation. 4. TRUE/FALSE of whether somatic regions passed QC
}
\description{
Calculate normalised coverage file for the predicted IGH germline haplotype for a tumour sample
}
