% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGH_somatic_norm_for_region.R
\name{IGH_somatic_norm_for_region}
\alias{IGH_somatic_norm_for_region}
\title{Function to normalise IGH coverage values by predicted somatic copy number alterations}
\usage{
IGH_somatic_norm_for_region(input_cov, region_df, threshold_to_norm = 0.2)
}
\arguments{
\item{input_cov}{data frame of coverage values in IGH loci}

\item{region_df}{data frame of predicted copy number of IGH loci regions for normalisations}

\item{threshold_to_norm}{threshold for genomic regions to be normalised}
}
\value{
Updated coverage data frame normalised by predicted somatic copy number alterations
}
\description{
Function to normalise IGH coverage values by predicted somatic copy number alterations
}
