% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuneLENS-package.R
\docType{package}
\name{ImmuneLENS-package}
\alias{ImmuneLENS}
\alias{ImmuneLENS-package}
\title{ImmuneLENS: Immune Lymphocyte Estimation from Nucleotide Sequencing}
\description{
Immune Lymphocyte Estimation from Nucleotide Sequencing (Immune LENS) is an R package to calculate T and B cell fractions from WGS or WES data from hg19 or hg38 aligned genomes. Includes functions from TcellExTRECT package for support of WES analysis.
}
\keyword{internal}
