% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImmuneLENS.R
\name{plotImmuneLENS}
\alias{plotImmuneLENS}
\title{Function to plot output from ImmuneLENS}
\usage{
plotImmuneLENS(
  vdj.region.df,
  vdj.gene = "TCRA",
  hg19_or_38 = "hg38",
  GC_correct = TRUE,
  GC_mode = "simultaneous",
  median.k = 50,
  median.thresh = 15,
  sample_name = "test",
  num.points = 3000,
  ylims = NULL,
  element.t.size = NULL,
  point.size = 0.5,
  customSeg = NULL,
  customFasta = NULL,
  customFlag = NULL,
  restriktor.absval = 1e-05,
  removed_flag = TRUE,
  output_df = FALSE,
  classSwitch = TRUE,
  customNorm = NULL,
  IGH_gc_constraint = TRUE,
  TCRB_gc_constraint = TRUE,
  IGH_gc_constraint_value = 0.01,
  allGC = FALSE
)
}
\arguments{
\item{vdj.region.df}{data frame containing coverage values by position}

\item{vdj.gene}{VDJ gene to be plotted}

\item{hg19_or_38}{hg19 or hg38 version of genome}

\item{GC_correct}{TRUE or FALSE, whether to correct for GC content}

\item{GC_mode}{GC correct running 'simultaneous' to linear model or 'prior'}

\item{median.k}{rolling median window}

\item{median.thresh}{threshold to remove exons with low coverage}

\item{sample_name}{name of sample run}

\item{num.points}{Number of data points}

\item{ylims}{limits for y axis}

\item{element.t.size}{size of text}

\item{point.size}{size of text}

\item{customSeg}{Custom df defining normalisation segments}

\item{customFasta}{Custom fasta file}

\item{customFlag}{Custom list of flagged exons to remove}

\item{restriktor.absval}{parameter for fitting model with restriktor}

\item{removed_flag}{Whether to removed flagged WGS locations (default = TRUE)}

\item{output_df}{If true return data frame used for plotting instead of plot}

\item{classSwitch}{Whether to include class switching in model for IGH (default = TRUE)}

\item{customNorm}{Custom dataframe with average values to normalise for, summarised over 100bp (default = NULL)}

\item{IGH_gc_constraint}{TRUE/FALSE, use additional constraints for GC correction smooth.gc and smooth.gc2 in the IGH solution (default TRUE to prevent over-fitting)}

\item{TCRB_gc_constraint}{TRUE/FALSE, use additional constraints for GC correction smooth.gc and smooth.gc2 in the TCRB solution (default TRUE to prevent over-fitting)}

\item{IGH_gc_constraint_value}{Value to constrain IGH solution (default 0.01)}

\item{allGC}{TRUE/FALSE, when TRUE make model where everything is due to GC}
}
\value{
data frame of TCRA T cell fractions with 95% CI
}
\description{
Function to plot output from ImmuneLENS
}
