% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSegmentBubble.R
\name{plotSegmentBubble}
\alias{plotSegmentBubble}
\title{Function to visualise segment output as bubble plot}
\usage{
plotSegmentBubble(segment_out, V_or_J, text_size = 6)
}
\arguments{
\item{segment_out}{data frame output from runTcellExtrect_WGS containing segment fractions}

\item{V_or_J}{plot V or J segment proportions}

\item{text_size}{size of text to plot in bubbles}
}
\value{
Bubble plots
}
\description{
Function to visualise segment output as bubble plot
}
