% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_somatic_correction_test.R
\name{run_somatic_correction_test}
\alias{run_somatic_correction_test}
\title{QC test for predicted somatic copy number alterations in IGH loci (max logR change > 0.1 and kb change length > 100)}
\usage{
run_somatic_correction_test(somatic_regions_df)
}
\arguments{
\item{somatic_regions_df}{data frame of predicted somatic copy number alterations in IGH loci}
}
\value{
TRUE or FALSE based on QC test
}
\description{
QC test for predicted somatic copy number alterations in IGH loci (max logR change > 0.1 and kb change length > 100)
}
