% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runImmuneLENS.R
\name{runImmuneLENS}
\alias{runImmuneLENS}
\title{Function to run ImmuneLENS}
\usage{
runImmuneLENS(
  vdj.region.df,
  vdj.gene = "TCRA",
  hg19_or_38 = "hg38",
  GC_correct = TRUE,
  median.k = 50,
  median.thresh = 15,
  sample_name = "test",
  customSeg = NULL,
  customFasta = NULL,
  customFlag = NULL,
  restriktor.absval = 1e-05,
  removed_flag = TRUE,
  GC_mode = "simultaneous",
  classSwitch_input = TRUE,
  customNorm = NULL
)
}
\arguments{
\item{vdj.region.df}{data frame containing coverage values by position}

\item{vdj.gene}{VDJ gene to use}

\item{hg19_or_38}{hg19 or hg38 version of genome}

\item{GC_correct}{whether to use GC correction or not for output}

\item{median.k}{rolling median window}

\item{median.thresh}{threshold to remove exons with low coverage}

\item{sample_name}{name of sample run}

\item{customSeg}{Custom df defining normalisation segments}

\item{customFasta}{Custom fasta file}

\item{customFlag}{Custom list of flagged exons to remove}

\item{restriktor.absval}{parameter for fitting model with restriktor}

\item{removed_flag}{Whether to removed flagged WGS locations (default = TRUE)}

\item{GC_mode}{Run GC correct 'simultaneous' or 'prior' to fitting linear model}

\item{classSwitch_input}{Whether to include class switching in model for IGH (default = TRUE)}

\item{customNorm}{Custom dataframe with average values to normalise for, summarised over 100bp (default = NULL)}
}
\value{
data frame of TCRA T cell fractions with 95\% CI
}
\description{
Function to run ImmuneLENS
}
