% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somatic_cna_IGH_caller.R
\name{somatic_cna_IGH_caller}
\alias{somatic_cna_IGH_caller}
\title{Calculate predicted regions containing somatic copy number alterations in IGH locus}
\usage{
somatic_cna_IGH_caller(
  germline_cov,
  tumour_cov,
  gc_correct = TRUE,
  run_breakpoint_check = FALSE,
  rpart_cp = 0.01
)
}
\arguments{
\item{germline_cov}{data frame of positions and coverage values for IGH loci of germline sample}

\item{tumour_cov}{data frame of positions and coverage values for IGH loci of tumour sample}

\item{gc_correct}{logical, should the normalisation be GC corrected? (default = TRUE)}

\item{run_breakpoint_check}{logical, should the normalisation be checked for breakpoints? (default = TRUE)}

\item{rpart_cp}{complexity parameter for rpart model (default = 0.01)}
}
\value{
Data frame containing regions of predicted somatic copy number alterations
}
\description{
Calculate predicted regions containing somatic copy number alterations in IGH locus
}
