% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseWGS.R
\name{summariseWGS}
\alias{summariseWGS}
\title{Function to summarise raw output from WGS version}
\usage{
summariseWGS(segmodel.out, vdj.gene = "TCRA", customNorm = NULL)
}
\arguments{
\item{segmodel.out}{output from getVDJfraction_WGS_segmodel}

\item{vdj.gene}{VDJ gene to use}

\item{customNorm}{Custom dataframe with average values to normalise for, summarised over 100bp (default = NULL)}
}
\description{
Function to summarise raw output from WGS version
}
