% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcra_seg_hg38_vdj}
\alias{tcra_seg_hg38_vdj}
\title{TCRA segments file with V/J segment loci (hg38)}
\format{
A list containing regions used to run T cell ExTRECT 2.0
}
\usage{
tcra_seg_hg38_vdj
}
\value{
NA
}
\description{
Segments used for running T cell ExTRECT, defining TCRA loci,
local regions used for normalisation and region of maximum VDJ
recombination and V/J segments (hg19 version)
}
\keyword{datasets}
