% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vdj.segments.list}
\alias{vdj.segments.list}
\title{TCRB/TCRG/IGH/IGL/IGJ segments file with V/J segment loci}
\format{
A list containing regions used to run T cell ExTRECT 2.0
}
\usage{
vdj.segments.list
}
\value{
NA
}
\description{
Segments used for running T cell ExTRECT, defining V/D/J segment
locations for both hg19 and hg38
}
\keyword{datasets}
