% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vdj_seg_list}
\alias{vdj_seg_list}
\title{VDJ segments file (hg19 and hg38)}
\format{
A list containing regions used to run T cell ExTRECT
}
\usage{
vdj_seg_list
}
\value{
NA
}
\description{
Segments used for running T cell ExTRECT, defining all VDJ genes loci,
local regions used for normalisation and region of maximum VDJ
recombination
}
\keyword{datasets}
