% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runImmuneLENS.R
\name{runImmuneLENS}
\alias{runImmuneLENS}
\title{Function to run ImmuneLENS}
\usage{
runImmuneLENS(
  vdj.region.df,
  vdj.gene = "TCRA",
  hg19_or_38 = "hg38",
  GC_correct = TRUE,
  median.k = 50,
  median.thresh = 15,
  sample_name = "test",
  customSeg = NULL,
  customFlag = NULL,
  removed_flag = TRUE,
  IGH_gc_constraint = TRUE,
  TCRB_gc_constraint = TRUE,
  IGH_gc_constraint_value = 0.01,
  allGC = FALSE
)
}
\arguments{
\item{vdj.region.df}{data frame containing coverage values by position}

\item{vdj.gene}{VDJ gene to use}

\item{hg19_or_38}{hg19 or hg38 version of genome}

\item{GC_correct}{whether to use GC correction or not for output}

\item{median.k}{rolling median window}

\item{median.thresh}{threshold to remove exons with low coverage}

\item{sample_name}{name of sample run}

\item{customSeg}{Custom df defining normalisation segments}

\item{customFlag}{Custom list of flagged exons to remove}

\item{removed_flag}{Whether to removed flagged WGS locations (default = TRUE)}

\item{IGH_gc_constraint}{TRUE/FALSE, use additional constraints for GC correction smooth.gc and smooth.gc2 in the IGH solution (default TRUE to prevent over-fitting)}

\item{TCRB_gc_constraint}{TRUE/FALSE, use additional constraints for GC correction smooth.gc and smooth.gc2 in the TCRB solution (default TRUE to prevent over-fitting)}

\item{IGH_gc_constraint_value}{Value to constrain IGH solution (default 0.01)}

\item{allGC}{TRUE/FALSE, when TRUE make model where everything is due to GC}
}
\value{
data frame of TCRA T cell fractions with 95\% CI
}
\description{
Function to run ImmuneLENS
}
